did some edits to make it usable

+++ student-distrib/isr.c
@@ -212,25 +212,25 @@ void isr18()
 
 /* Set first 32 entries in IDT to Intel's defined exceptions */
 void isrs_install(){
-	set_trap_gate(0,  isr0);
-	set_trap_gate(1,  isr1);
-	set_trap_gate(2,  isr2);
-	set_trap_gate(3,  isr3);
-	set_trap_gate(4,  isr4);
-	set_trap_gate(5,  isr5);
-	set_trap_gate(6,  isr6);
-	set_trap_gate(7,  isr7);
-	set_trap_gate(8,  isr8);
-	set_trap_gate(9,  isr9);
-	set_trap_gate(10,  isr10);
-	set_trap_gate(11,  isr11);
-	set_trap_gate(12,  isr12);
-	set_trap_gate(13,  isr13);
-	set_trap_gate(14,  isr14);
-	set_trap_gate(15,  isr15);
-	set_trap_gate(16,  isr16);
-	set_trap_gate(17,  isr17);
-	set_trap_gate(18,  isr18);
+	set_trap_gate(0,  (unsigned long)isr0);
+	set_trap_gate(1,  (unsigned long)isr1);
+	set_trap_gate(2,  (unsigned long)isr2);
+	set_trap_gate(3,  (unsigned long)isr3);
+	set_trap_gate(4,  (unsigned long)isr4);
+	set_trap_gate(5,  (unsigned long)isr5);
+	set_trap_gate(6,  (unsigned long)isr6);
+	set_trap_gate(7,  (unsigned long)isr7);
+	set_trap_gate(8,  (unsigned long)isr8);
+	set_trap_gate(9,  (unsigned long)isr9);
+	set_trap_gate(10,  (unsigned long)isr10);
+	set_trap_gate(11,  (unsigned long)isr11);
+	set_trap_gate(12,  (unsigned long)isr12);
+	set_trap_gate(13,  (unsigned long)isr13);
+	set_trap_gate(14,  (unsigned long)isr14);
+	set_trap_gate(15,  (unsigned long)isr15);
+	set_trap_gate(16,  (unsigned long)isr16);
+	set_trap_gate(17,  (unsigned long)isr17);
+	set_trap_gate(18,  (unsigned long)isr18);
 }
